ui tweaks for p.mgmt, Create PerspectiveEditor, PerspectiveCaption module to contain different aspects
