+++ graphcap_studio/src/features/perspectives/components/PerspectiveCaption/PerspectiveActions/GenerateAllButton.tsx
