+++ graphcap_studio/src/features/perspectives/components/PerspectiveCaption/PerspectiveCard/MetadataDisplay.tsx
