+++ graphcap_studio/src/features/perspectives/components/PerspectiveCaption/PerspectiveCard/debug-fields/Separator.tsx
