+++ graphcap_studio/src/features/perspectives/components/PerspectiveCaption/PerspectiveCard/schema-fields/CaptionContext.tsx
