+++ graphcap_studio/src/features/perspectives/components/PerspectiveCaption/PerspectiveNavigation/PerspectiveHeader.tsx
