+++ graphcap_studio/src/features/perspectives/components/PerspectiveCaption/PerspectiveNavigation/PerspectivesPager.tsx
