+++ graphcap_studio/src/features/perspectives/components/PerspectiveManagement/ErrorDisplay.tsx
