+++ graphcap_studio/src/features/perspectives/components/PerspectiveManagement/PerspectiveEditor/index.ts
@@ -1,0 +1,1 @@
+export { PerspectiveEditor } from './PerspectiveEditor';
