+++ graphcap_studio/src/features/perspectives/components/PerspectiveManagement/PerspectiveModules/ModuleInfo.tsx
