--- graphcap_studio/src/features/perspectives/components/PerspectiveModules/index.ts
@@ -1,6 +1,0 @@
-export { ModuleList } from './ModuleList';
-export { ModuleInfo } from './ModuleInfo';
-export { NotFound } from './NotFound';
-export { ErrorDisplay } from './ErrorDisplay';
-export { LoadingDisplay } from './LoadingDisplay';
-export { PerspectiveDetail } from './PerspectiveDetail';
