+++ graphcap_studio/src/pages/perspectives/index.ts
@@ -1,0 +1,2 @@
+export { ModulePage } from './module/ModulePage';
+export { PerspectiveDetailPage } from './module/PerspectiveDetailPage';
