+++ graphcap_studio/src/pages/perspectives/module/index.ts
@@ -1,0 +1,2 @@
+export { ModulePage } from './ModulePage';
+export { PerspectiveDetailPage } from './PerspectiveDetailPage';
